Fix MakeLenses
The `trash` variable was used twice, which caused problems trying to unify `FormSpec` and `InputField` for obvious reasons; just removing the `row` variable fixed that.

The other problem was that the `form` variable was not solved, meaning its constraints could not be resolved. Passing it to the class fixes this.